+++ src/dnscrypt-proxy/pid_file.c
@@ -97,7 +97,8 @@ pid_file_write(const int fd, const pid_t child)
     pid_buf_len = evutil_snprintf(pid_buf, sizeof pid_buf, "%llu",
                                   (unsigned long long) child);
     assert((size_t) pid_buf_len < sizeof pid_buf);
-    if (safe_write(fd, pid_buf, pid_buf_len, -1) != pid_buf_len) {
+    if (safe_write(fd, pid_buf, (size_t) pid_buf_len, -1) !=
+        (ssize_t) pid_buf_len) {
         (void) ftruncate(fd, (off_t) 0);
         (void) close(fd);
         return -1;
